+++ src/Http/Middleware/Authenticate.php
@@ -1,0 +1,31 @@
+<?php
+
+/*
+ * This file is part of Laravel Sense.
+ *
+ * (c) Anton Komarev <a.komarev@cybercog.su>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Cog\Laravel\Sense\Http\Middleware;
+
+use Cog\Laravel\Sense\Authentication\Services\Authenticator;
+
+class Authenticate
+{
+    /**
+     * Handle the incoming request.
+     *
+     * @param  \Illuminate\Http\Request $request
+     * @param  \Closure $next
+     * @return \Illuminate\Http\Response|null
+     */
+    public function handle($request, $next)
+    {
+        return Authenticator::check($request) ? $next($request) : abort(403);
+    }
+}