+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # Changelog
 
-## 1.0.0 - 2018-09-25
+All notable changes to `laravel-sense` will be documented in this file.
+
+## [0.2.0] - 2018-09-26
+
+### Added
+
+- ([#9](https://github.com/cybercog/laravel-sense/pull/9)) Dashboard Authenticator
+
+## 0.1.0 - 2018-09-25
 
 - Initial release
 
-All notable changes to `laravel-sense` will be documented in this file.
+[0.2.0]: https://github.com/cybercog/laravel-sense/compare/0.1.0...0.2.0